refactored more c#
