+++ src/ai/.x/templates/openai-asst-cs/_.json
@@ -9,7 +9,7 @@
 
   "_IS_BROWSER_TEMPLATE": "false",
   "_IS_OPENAI_ASST_TEMPLATE": "true",
-  "_IS_OPENAI_ASST_STREAMING_TEMPLATE": "true",
+  "_IS_OPENAI_ASST_STREAMING_TEMPLATE": "false",
   "_IS_OPENAI_ASST_CODE_INTERPRETER_TEMPLATE": "false",
   "_IS_OPENAI_ASST_FILE_SEARCH_TEMPLATE": "false",
   "_IS_OPENAI_ASST_FUNCTIONS_TEMPLATE": "false"
