+++ src/ai/.x/templates/openai-asst-streaming-with-file-search-cs/_.json
@@ -11,6 +11,6 @@
   "_IS_OPENAI_ASST_TEMPLATE": "true",
   "_IS_OPENAI_ASST_STREAMING_TEMPLATE": "true",
   "_IS_OPENAI_ASST_CODE_INTERPRETER_TEMPLATE": "false",
-  "_IS_OPENAI_ASST_FILE_SEARCH_TEMPLATE": "false",
+  "_IS_OPENAI_ASST_FILE_SEARCH_TEMPLATE": "true",
   "_IS_OPENAI_ASST_FUNCTIONS_TEMPLATE": "false"
 }
