+++ src/ai/.x/templates/openai-asst-streaming-with-functions-cs/_.json
@@ -12,5 +12,5 @@
   "_IS_OPENAI_ASST_STREAMING_TEMPLATE": "true",
   "_IS_OPENAI_ASST_CODE_INTERPRETER_TEMPLATE": "false",
   "_IS_OPENAI_ASST_FILE_SEARCH_TEMPLATE": "false",
-  "_IS_OPENAI_ASST_FUNCTIONS_TEMPLATE": "false"
+  "_IS_OPENAI_ASST_FUNCTIONS_TEMPLATE": "true"
 }
