+++ src/ai/.x/templates/speech-to-text-with-file-cs/_.json
@@ -1,7 +1,11 @@
 {
-  "_LongName": "Speech-to-text (w/ File input)",
-  "_ShortName": "speech-to-text-with-file",
-  "_Language": "C#",
-  "AZURE_AI_SPEECH_KEY": "<insert your Speech Service API key here>",
-  "AZURE_AI_SPEECH_REGION": "<insert your Speech Service region here>"
+    "_LongName": "Speech-to-text (w/ File input)",
+    "_ShortName": "speech-to-text-with-file",
+    "_Language": "C#",
+    "AZURE_AI_SPEECH_KEY": "<insert your Speech Service API key here>",
+    "AZURE_AI_SPEECH_REGION": "<insert your Speech Service region here>",
+    "_IS_SPEECH_TO_TEXT_CONTINUOUS": true,
+    "_IS_SPEECH_TO_TEXT_WITH_FILE": true,
+    "_IS_SPEECH_TO_TEXT_WITH_KEYWORD": false,
+    "_IS_SPEECH_TO_TEXT_WITH_TRANSLATION": false
 }
