+++ src/ai/.x/templates/speech-to-text-with-keyword-cs/_.json
@@ -1,7 +1,11 @@
 {
-  "_LongName": "Speech-to-text (w/ Keyword detection)",
-  "_ShortName": "speech-to-text-with-keyword",
-  "_Language": "C#",
-  "AZURE_AI_SPEECH_KEY": "<insert your Speech Service API key here>",
-  "AZURE_AI_SPEECH_REGION": "<insert your Speech Service region here>"
+    "_LongName": "Speech-to-text (w/ Keyword detection)",
+    "_ShortName": "speech-to-text-with-keyword",
+    "_Language": "C#",
+    "AZURE_AI_SPEECH_KEY": "<insert your Speech Service API key here>",
+    "AZURE_AI_SPEECH_REGION": "<insert your Speech Service region here>",
+    "_IS_SPEECH_TO_TEXT_CONTINUOUS": true,
+    "_IS_SPEECH_TO_TEXT_WITH_FILE": false,
+    "_IS_SPEECH_TO_TEXT_WITH_KEYWORD": true,
+    "_IS_SPEECH_TO_TEXT_WITH_TRANSLATION": false
 }
