+++ src/ai/.x/templates/text-to-speech-with-file-cs/TextToSpeechWithFile.csproj._
@@ -1,15 +1,1 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <EnableDefaultCompileItems>true</EnableDefaultCompileItems>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CognitiveServices.Speech" Version="1.35.0" />
-  </ItemGroup>
-
-</Project>
+{{@include text-to-speech-cs/TextToSpeech.csproj._}}
